requirements.txt
pandas==2.2.2
pydantic==2.8.2
requests==2.32.3
Pillow==10.4.0
tqdm==4.66.4
python-dotenv==1.0.1
openpyxl==3.1.5
rich==13.7.1
typer==0.12.3
PyYAML==6.0.2
faiss-cpu==1.8.0.post1
open-clip-torch==2.24.0
sentence-transformers==3.0.1
scikit-learn==1.5.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
gradio==4.44.1
